refactor(lint): move isHookInput to hooks/lint.ts
Hook-specific logic belongs in hooks/, lint logic in scripts/.
