+++ scripts/lint.ts
@@ -1,25 +1,7 @@
 import { dirname, join } from "node:path";
 
-interface HookInput {
-	// eslint-disable-next-line flawless/naming-convention -- Upstream-defined input shape
-	tool_input: {
-		// eslint-disable-next-line flawless/naming-convention -- Upstream-defined input shape
-		file_path: string;
-	};
-}
-
 interface FileSystemDeps {
 	existsSync(path: string): boolean;
-}
-
-export function isHookInput(value: unknown): value is HookInput {
-	if (typeof value !== "object" || value === null || !("tool_input" in value)) {
-		return false;
-	}
-
-	// eslint-disable-next-line flawless/naming-convention -- Upstream-defined input shape
-	const { tool_input } = value as { tool_input: unknown };
-	return typeof tool_input === "object" && tool_input !== null && "file_path" in tool_input;
 }
 
 const DEFAULT_EXTENSIONS = [".ts", ".tsx", ".js", ".jsx", ".mjs", ".mts"];
